Increase player base for colissions
Previously the point of contact between the player
and a platform was 1 pixel width. This caused the
impression that the ball slipped through a platform
when it should have actually collided with it, in
cases where the player just missed the edge of a platform.

Increase this colission width from 1 pixel to a higher,
configurable value to make these scenarios feel more natural.